requirements.txt
# ── Core ML / NLP ──────────────────────────────────────────────────────────
sentence-transformers==3.4.1
faiss-cpu==1.10.0
scikit-learn==1.6.1
numpy==2.2.3

# ── PDF Parsing ─────────────────────────────────────────────────────────────
pymupdf

# ── Text Splitting ───────────────────────────────────────────────────────────
langchain-text-splitters

# ── Keyword Extraction ───────────────────────────────────────────────────────
keybert

# ── LLM — Groq (cloud inference) ─────────────────────────────────────────────
groq

# ── Environment / Config ─────────────────────────────────────────────────────
python-dotenv

# ── PDF Generation ──────────────────────────────────────────────────────────
fpdf2

# ── Streamlit App ────────────────────────────────────────────────────────────
streamlit
pandas
